ui: 장르버튼 컴포넌트 통합, ui 디테일 - 컬러 변수 추가
