+++ src/components/common/GenreButtons/GenreButtons.module.scss
@@ -1,0 +1,20 @@
+@import "../../../styles/_variables";
+@import "../../../styles/_mixin";
+
+.category_nav {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  row-gap: 8px;
+  column-gap: 12px;
+
+  &.main {
+    margin-top: 36px;
+  }
+
+  &.search {
+    margin: 16px 0 24px 0;
+    border-top: 1px solid $line-light;
+    border-bottom: 1px solid $line-light;
+    padding: 20px 0;
+  }
+}
