+++ src/components/common/StarTag/StarTag.module.scss
@@ -5,7 +5,7 @@
   display: flex;
   width: fit-content;
   align-items: center;
-  background-color: $primary-light;
+  background-color: $primary-lighter;
   padding: 5px 6px;
   border-radius: 4px;
   font-size: $font-md;
