+++ src/styles/_variables.scss
@@ -8,7 +8,8 @@ $white-light: #ffffffcc;
 $white-lighter: #ffffff99;
 
 $primary: #7926ff;
-$primary-light: #7926ff1f;
+$primary-light: #7926ff66;
+$primary-lighter: #7926ff1f;
 
 $danger: #ed4b4b;
 $danger-light: #ed4b4b1f;
